ai_image_enhance/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pillow==10.4.0
numpy==1.26.4
opencv-python==4.8.1.78
torch==2.2.2
torchvision==0.17.2
pydantic==2.8.2
python-dotenv==1.0.1
